Error handling during login

+++ commands/delete_customer_linker.py
@@ -11,13 +11,14 @@ console = Console()
 
 def delete_customer_linker_command(jobId):
     token = get_token()
-    apiurl = f'{config.api_base_url}/job/customer-linker/{jobId}'
-    jobs = requests.delete(apiurl, headers={'Authorization': f'Bearer {token}'})
-    if jobs.status_code == 200:
-        console.print(
-            f"[bold green]Job Deleted if it exists.  [/bold "
-            f"green]")
-    else:
-        console.print(f"[bold red]Invalid Access Token. Try logging in again and re-run the same command again [/bold "
-                      f"red]")
-        common_login()
+    if len(token) > 0:
+        apiurl = f'{config.api_base_url}/job/customer-linker/{jobId}'
+        jobs = requests.delete(apiurl, headers={'Authorization': f'Bearer {token}'})
+        if jobs.status_code == 200:
+            console.print(
+                f"[bold green]Job Deleted if it exists.  [/bold "
+                f"green]")
+        else:
+            console.print(f"[bold red]Invalid Access Token. Try logging in again and re-run the same command again [/bold "
+                          f"red]")
+            common_login()
